vision/requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.1
opencv-python==4.5.1.48
pymata4==1.10
pyserial==3.5
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
redis==3.5.3
requests==2.25.1
urllib3==1.26.3
